+++ app/src/main/java/io/github/muntashirakon/AppManager/AppManager.java
@@ -18,6 +18,7 @@ import org.lsposed.hiddenapibypass.HiddenApiBypass;
 import java.security.Security;
 
 import dalvik.system.ZipPathValidator;
+import io.github.muntashirakon.AppManager.debloat.DebloatDefinitionsUpdater;
 import io.github.muntashirakon.AppManager.misc.AMExceptionHandler;
 import io.github.muntashirakon.AppManager.misc.ProfilingTriggerHelper;
 import io.github.muntashirakon.AppManager.utils.Utils;
@@ -46,6 +47,7 @@ public class AppManager extends Application {
         Security.addProvider(new JavaKeyStoreProvider());
         Security.addProvider(new BouncyCastleProvider());
         ProfilingTriggerHelper.registerTriggersIfSupported(this);
+        DebloatDefinitionsUpdater.scheduleUpdateIfAllowed(this);
     }
 
     private void configureActivityEmbeddingSplits() {